+++ src/Repositories/BaseRepository.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace SleepingOwl\Admin\Repository;
+namespace SleepingOwl\Admin\Repositories;
 
 use Illuminate\Database\Eloquent\Model;
-use SleepingOwl\Admin\Contracts\RepositoryInterface;
+use SleepingOwl\Admin\Contracts\Repositories\RepositoryInterface;
 use SleepingOwl\Admin\Exceptions\RepositoryException;
 
 class BaseRepository implements RepositoryInterface