+++ tests/Admin/Form/FormDefaultTest.php
@@ -2,7 +2,7 @@
 
 use Mockery as m;
 use SleepingOwl\Admin\Form\FormDefault;
-use SleepingOwl\Admin\Contracts\RepositoryInterface;
+use SleepingOwl\Admin\Contracts\Repositories\RepositoryInterface;
 use SleepingOwl\Admin\Contracts\Form\FormButtonsInterface;
 use SleepingOwl\Admin\Contracts\Form\FormElementInterface;
 use SleepingOwl\Admin\Contracts\ModelConfigurationInterface;